requirements.txt
streamlit>=1.28.1
selenium>=4.10.0
anthropic>=0.7.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
requests>=2.31.0
modal>=0.55.0
